% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alternative_forms.R
\name{get_alternative_forms}
\alias{get_alternative_forms}
\title{get alternative_forms of molecules}
\usage{
get_alternative_forms(ChEMBLIDs, verbose = FALSE)
}
\arguments{
\item{ChEMBLIDs}{a vector of ChEMBL IDs, such as c("CHEMBL4303288", "CHEMBL1380")}

\item{verbose}{verbose.}
}
\value{
a tibble object with 3 columns, molecule_chembl_id, parent_chembl_id and alternative_forms.
}
\description{
ChEMBL includes parent molecules and their salts (approved and investigational)
as well as other alternative forms such as hydrates and radioisotopes.
These alternative forms are linked to their parent compound through the
molecule hierarchy.
ref: http://chembl.blogspot.com/2020/09/molecule-hierarchy.html
}
\examples{
ChEMBLIDs <- c("CHEMBL4303288", "CHEMBL1380", "CHEMBL3040746")
res_tib <- get_alternative_forms(ChEMBLIDs)
}
