% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inds.R
\name{get_inds}
\alias{get_inds}
\title{Get Drug indications based on ChEMBL IDs}
\usage{
get_inds(ChEMBLIDs, verbose = FALSE)
}
\arguments{
\item{ChEMBLIDs}{a vector of ChEMBL IDs, such as c("CHEMBL4303288", "CHEMBL1380")}

\item{verbose}{verbose.}
}
\value{
a tibble object with 3 columns, consisting of molecule_chembl_id, mesh_heading, efo_term.
}
\description{
Get Drug indications based on ChEMBL IDs
}
\examples{

ind_tib <- get_inds(c("CHEMBL4303288", "CHEMBL1380"))
dim(ind_tib)


}
