% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MOA.R
\name{get_MOA}
\alias{get_MOA}
\title{Get Drug Mechanism and Targets based on ChEMBL IDs}
\usage{
get_MOA(ChEMBLIDs, full = FALSE, verbose = FALSE)
}
\arguments{
\item{ChEMBLIDs}{a vector of ChEMBL IDs, such as c("CHEMBL4303288", "CHEMBL1380")}

\item{full}{FASLE (default) or TRUE. outputs molecule_chembl_id, molecular_mechanism, target or all the related information}

\item{verbose}{verbose.}
}
\value{
a tibble object with 3 columns, consisting of molecule_chembl_id, molecular_mechanism, target.
}
\description{
Get Drug Mechanism and Targets based on ChEMBL IDs
}
\examples{
ChEMBLIDs <- c("CHEMBL4303288", "CHEMBL1380")
res_tib <- get_MOA(ChEMBLIDs)
dim(res_tib)


}
