% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parent_chembl_id.R
\name{get_parent_chembl_id}
\alias{get_parent_chembl_id}
\title{get parent chembl ID}
\usage{
get_parent_chembl_id(ChEMBLIDs)
}
\arguments{
\item{ChEMBLIDs}{a vector of ChEMBL IDs, such as c("CHEMBL4303288", "CHEMBL1380")}
}
\value{
a tibble object with 2 columns, molecule_chembl_id and parent_chembl_id
}
\description{
ChEMBL includes parent molecules and their salts (approved and investigational)
as well as other alternative forms such as hydrates and radioisotopes.
These alternative forms are linked to their parent compound through the
molecule hierarchy.
ref: http://chembl.blogspot.com/2020/09/molecule-hierarchy.html
}
\examples{
ChEMBLIDs <- c("CHEMBL4303288", "CHEMBL1380")
parent_CHEMBLID_tib <- get_parent_chembl_id(ChEMBLIDs)
dim(parent_CHEMBLID_tib)
}
