% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_properties.R
\name{get_properties}
\alias{get_properties}
\title{get_properties}
\usage{
get_properties(ChEMBLIDs, verbose = FALSE)
}
\arguments{
\item{ChEMBLIDs}{a vector of ChEMBL IDs, such as c("CHEMBL4303288", "CHEMBL1380")}

\item{verbose}{verbose.
#' @import dplyr}
}
\description{
query molecule information, including properties, structural representations and synonyms of drugs.
}
\examples{
ChEMBLIDs <- c("CHEMBL4303288", "CHEMBL1380", "CHEMBL345714")
res_tib <- get_properties(ChEMBLIDs)
dim(res_tib)
}
