% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pasteX}
\alias{pasteX}
\title{pasteX}
\usage{
pasteX(vec, collapse = ";")
}
\arguments{
\item{vec}{a char vector.}

\item{collapse}{which collapse is to used in `paste`.}
}
\value{
a string
}
\description{
paste a vector to a string (for dplyr summarize chars).
}
